+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/chat/ExtendedAbstracts.kt
@@ -1,7 +1,6 @@
 package dev.inmo.tgbotapi.types.chat
 
 import dev.inmo.tgbotapi.types.*
-import dev.inmo.tgbotapi.types.chat.ExtendedChat
 import dev.inmo.tgbotapi.types.message.abstracts.Message
 import dev.inmo.tgbotapi.types.message.abstracts.TelegramBotAPIMessageDeserializeOnlySerializer
 import kotlinx.serialization.Serializable
@@ -39,6 +38,16 @@ sealed interface ExtendedSupergroupChat : SupergroupChat, ExtendedGroupChat {
     val canSetStickerSet: Boolean
     val linkedChannelChatId: ChatId?
     val location: ChatLocation?
+
+    /**
+     * This field represents field "join_to_send_messages" from API
+     */
+    val requiresJoinForMessaging: Boolean
+
+    /**
+     * This field represents field "join_by_request" from API
+     */
+    val requireAdminApproveToJoin: Boolean
 }
 
 @Serializable(ExtendedChatSerializer::class)