+++ tgbotapi.webapps/src/jsMain/kotlin/dev/inmo/tgbotapi/webapps/BackButton.kt
@@ -1,0 +1,11 @@
+package dev.inmo.tgbotapi.webapps
+
+external interface BackButton {
+    val isVisible: Boolean
+
+    fun onClick(callback: () -> Unit)
+    fun offClick(callback: () -> Unit)
+
+    fun show()
+    fun hide()
+}