+++ tgbotapi.webapps/src/jsMain/kotlin/dev/inmo/tgbotapi/webapps/EventType.kt
@@ -4,4 +4,7 @@ sealed class EventType(val typeName: String) {
     object ThemeChanged : EventType("themeChanged")
     object ViewportChanged : EventType("viewportChanged")
     object MainButtonClicked : EventType("mainButtonClicked")
+    object BackButtonClicked : EventType("backButtonClicked")
+    object SettingsButtonClicked : EventType("settingsButtonClicked")
+    object InvoiceClosed : EventType("invoiceClosed")
 }