+++ tgbotapi.webapps/src/jsMain/kotlin/dev/inmo/tgbotapi/webapps/MainButton.kt
@@ -23,6 +23,7 @@ external class MainButton {
     fun hideProgress(): MainButton
 
     internal fun onClick(eventHandler: () -> Unit): MainButton
+    internal fun offClick(eventHandler: () -> Unit): MainButton
 
     internal fun setParams(params: Json): MainButton
 }