+++ tgbotapi.webapps/src/jsMain/kotlin/dev/inmo/tgbotapi/webapps/ThemeParams.kt
@@ -2,15 +2,32 @@ package dev.inmo.tgbotapi.webapps
 
 external interface ThemeParams {
     @JsName("bg_color")
-    val backgroundColor: String?
+    val backgroundColor: HEXColor?
+    @JsName("secondary_bg_color")
+    val secondaryBackgroundColor: HEXColor?
     @JsName("text_color")
-    val textColor: String?
+    val textColor: HEXColor?
     @JsName("hint_color")
-    val hintColor: String?
+    val hintColor: HEXColor?
     @JsName("link_color")
-    val linkColor: String?
+    val linkColor: HEXColor?
     @JsName("button_color")
-    val buttonColor: String?
+    val buttonColor: HEXColor?
     @JsName("button_text_color")
-    val buttonTextColor: String?
+    val buttonTextColor: HEXColor?
+
+    @JsName("bg_color")
+    val backgroundColorHex: Color.Hex?
+    @JsName("secondary_bg_color")
+    val secondaryBackgroundColorHex: Color.Hex?
+    @JsName("text_color")
+    val textColorHex: Color.Hex?
+    @JsName("hint_color")
+    val hintColorHex: Color.Hex?
+    @JsName("link_color")
+    val linkColorHex: Color.Hex?
+    @JsName("button_color")
+    val buttonColorHex: Color.Hex?
+    @JsName("button_text_color")
+    val buttonTextColorHex: Color.Hex?
 }