+++ tgbotapi.webapps/src/jsMain/kotlin/dev/inmo/tgbotapi/webapps/haptic/HapticFeedback.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.webapps.haptic
+
+external interface HapticFeedback {
+    fun impactOccurred(style: HapticFeedbackStyle)
+    fun notificationOccurred(type: HapticFeedbackType)
+    fun selectionChanged()
+}