+++ tgbotapi.webapps/src/jsMain/kotlin/dev/inmo/tgbotapi/webapps/invoice/InvoiceClosedInfo.kt
@@ -1,0 +1,16 @@
+package dev.inmo.tgbotapi.webapps.invoice
+
+external interface InvoiceClosedInfo {
+    val url: String
+    val status: String
+}
+
+val InvoiceClosedInfo.statusTyped
+    get() = when (status) {
+        InvoiceStatus.Paid.name -> InvoiceStatus.Paid
+        InvoiceStatus.Cancelled.name -> InvoiceStatus.Cancelled
+        InvoiceStatus.Failed.name -> InvoiceStatus.Failed
+        InvoiceStatus.Pending.name -> InvoiceStatus.Pending
+        else -> InvoiceStatus.Unknown(status)
+    }
+